+++ docs/en/changes/changes.md
@@ -55,6 +55,7 @@
 * Register ValueColumn as Tag for Record in BanyanDB storage plugin.
 * Bump up Netty to 4.1.86.
 * Remove unnecessary additional columns when storage is in logical sharding mode.
+* The cluster coordinator support watch mechanism for notifying `RemoteClientManager` and `ServerStatusService`.
 * Fix ServiceMeshServiceDispatcher overwrite ServiceDispatcher debug file when open SW_OAL_ENGINE_DEBUG.
 
 #### UI